something approximating a POC for using Void for Custom
Obviously, this is extremely incomplete (really, one would want to
supply Void as the default type parameter (and use the `void`
package)), but it does allow one to do things like this:

```rust
    fn test_err() {
        let input = b"asdf";

        enum CustomErr {
            Foo,
            Bar
        }

        let err_test = add_return_error!(&input[..], ErrorKind::Custom(CustomErr::Foo),
                                         tag!(b"adf"));
        // still works with u32
        let err_test2 = add_return_error!(&input[..], ErrorKind::Custom(1u32),
                                         tag!(b"asdf"));
    }
```

and they Just Work.

Many many tests currently fail and I have no idea how involved it
would be to really do this thoroughly, but it seems both safer and
more correct than defaulting to u32: since none of the built-in
parsers should be using ErrorKind::Custom anyway, why not rule it out,
and guarantee that the errors they generate can always be converted to
an ErrorKind with E != Void?

The various trait tricks seem … not great, admittedly.